KhantoImoveis/app/apis/serializers/reservation_serializer.py
from rest_framework import serializers
from app.models import Reservation


class ReservationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Reservation
        fields = '__all__'
