requirements.txt
asgiref==3.6.0
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
Django==4.1.5
django-filter==22.1
djangorestframework==3.14.0
Faker==15.3.4
Markdown==3.4.1
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
toposort==1.7
tzdata==2022.7
wincertstore==0.2
